# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Tolk Translation API",
        "description": "OpenAPI documentation for the versioned Laravel API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List translations",
                "operationId": "listTranslations",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter by locale code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "mobile"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by translation key",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "welcome"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Search inside translation content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Welcome"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Free-text search across supported fields",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Welcome"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Success",
                                    "data": {
                                        "data": [
                                            {
                                                "id": 1,
                                                "translation_key": "welcome_message",
                                                "content": "Welcome",
                                                "locale": {
                                                    "id": 1,
                                                    "code": "en",
                                                    "name": "English"
                                                },
                                                "tags": [
                                                    {
                                                        "id": 1,
                                                        "name": "web"
                                                    }
                                                ],
                                                "created_at": "2026-06-24T10:00:00Z",
                                                "updated_at": "2026-06-24T10:00:00Z"
                                            }
                                        ],
                                        "links": {
                                            "first": null,
                                            "last": null,
                                            "prev": null,
                                            "next": null
                                        },
                                        "meta": {
                                            "current_page": 1,
                                            "from": 1,
                                            "last_page": 1,
                                            "path": "http://localhost/api/v1/translations",
                                            "per_page": 20,
                                            "to": 1,
                                            "total": 1
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a translation",
                "operationId": "storeTranslation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TranslationStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Created",
                                    "data": {
                                        "id": 1,
                                        "translation_key": "welcome_message",
                                        "content": "Welcome",
                                        "locale": {
                                            "id": 1,
                                            "code": "en",
                                            "name": "English"
                                        },
                                        "tags": [
                                            {
                                                "id": 1,
                                                "name": "web"
                                            }
                                        ],
                                        "created_at": "2026-06-24T10:00:00Z",
                                        "updated_at": "2026-06-24T10:00:00Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/translations/export/{locale}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations for a locale",
                "operationId": "exportTranslations",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locale export payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "welcome_message": "Welcome",
                                    "login": "Login"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SuccessResponse": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "id": 1
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "message": [
                                "Something went wrong."
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "required": [
                    "success",
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ],
                            "password": [
                                "The password field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}